refactor: refactor dyna writer to improve maintainability (#1101)
