+++ doc/source/conf.py
@@ -47,14 +47,9 @@ html_theme_options = {
     "ansys_sphinx_theme_autoapi": {
         "project": project,
         "ignore": [
-            "*writer*",
+            # ignore all files under the ``writer`` subpackage.
+            "*heart/writer/**/*",
             "*misc*",
-            "*custom_keywords*",
-            "*system_models.py",
-            "*define_function_templates.py",
-            "*heart_decks.py",
-            "*keyword_utils.py",
-            "*material_keywords.py",
         ],
     },
 }
