+++ src/ansys/health/heart/writer/custom_keywords/keywords/mat_077h.py
@@ -29,13 +29,13 @@ from ansys.dyna.core.lib.card import Card, Field
 from ansys.dyna.core.lib.keyword_base import KeywordBase
 
 """
-This files contains the keywords that is not supported by dynalib
+This files contains the keywords that is not supported by the PyDYNA keywords module
 """
 
 
 class Mat077H(KeywordBase):
     """DYNA MAT_077_H keyword
-    Replace the bug in current version of dynalib
+    Replace the bug in current version of PyDYNA keywords module
 
     """
